+++ tests/issues/issue-105/formatted.vue
@@ -1,0 +1,10 @@
+<template lang="pug">
+// intended not to be single quoted
+ComponentBar(foo="bar", bar="baz")
+</template>
+
+<script>
+  export default {
+    name: 'FooComponent' // intended to be single quote
+  }
+</script>