+++ tests/issues/issue-105/unformatted.vue
@@ -1,0 +1,13 @@
+<template lang="pug">
+// intended not to be single quoted
+ComponentBar(
+  foo="bar"
+  bar="baz"
+)
+</template>
+
+<script>
+  export default {
+    name: 'FooComponent' // intended to be single quote
+  }
+</script>